Create d_p_a/rope.py and move RoPE methods from  pytorch/attention.py to it
